requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
psycopg[binary]>=3.2,<3.3
django-environ>=0.11,<0.12
django-storages[s3]>=1.14,<1.15
redis>=5.0,<5.1
gunicorn>=22.0,<22.1
Pillow>=10.4,<10.5
whitenoise>=6.7,<7.0
